# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Test Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/test-runs:
    get:
      operationId: listTestRuns
      tags: [test-runs]
      summary: List test runs for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, passed, failed, aborted]
      responses:
        "200":
          description: List of test runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestRun"
    post:
      operationId: createTestRun
      tags: [test-runs]
      summary: Create a test run
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestRunBody"
      responses:
        "201":
          description: Created test run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRun"
  /projects/{projectId}/test-runs/{testRunId}:
    get:
      operationId: getTestRun
      tags: [test-runs]
      summary: Get a test run with results
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: testRunId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test run details with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRunDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTestRun
      tags: [test-runs]
      summary: Update a test run status
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: testRunId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestRunBody"
      responses:
        "200":
          description: Updated test run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRun"
    delete:
      operationId: deleteTestRun
      tags: [test-runs]
      summary: Delete a test run
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: testRunId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Test Results within a run
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    # Project schemas
...
    TestRun:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [pending, running, passed, failed, aborted]
        branch:
          type: string
          nullable: true
        commitHash:
          type: string
          nullable: true
        commitMessage:
          type: string
          nullable: true
        triggeredBy:
          type: string
          nullable: true
        ciProvider:
          type: string
          nullable: true
        totalTests:
          type: integer
        passedTests:
          type: integer
        failedTests:
          type: integer
        skippedTests:
          type: integer
        blockedTests:
          type: integer
        duration:
          type: integer
          nullable: true
          description: Duration in milliseconds
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - status
        - totalTests
        - passedTests
        - failedTests
        - skippedTests
        - blockedTests
        - createdAt
        - updatedAt
    TestRunDetail:
      allOf:
        - $ref: "#/components/schemas/TestRun"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/TestResult"
          required:
            - results
    CreateTestRunBody:
      type: object
      properties:
        name:
          type: string
        branch:
          type: string
        commitHash:
          type: string
        commitMessage:
          type: string
        triggeredBy:
          type: string
        ciProvider:
          type: string
        testCaseIds:
          type: array
          items:
            type: integer
      required:
        - name
    UpdateTestRunBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, passed, failed, aborted]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        duration:
          type: integer

    # Test result schemas
    TestResult:
      type: object
      properties:
        id:
          type: integer
        testRunId:
          type: integer
        testCaseId:
          type: integer
        testCaseTitle:
          type: string
        status:
          type: string
          enum: [pending, passed, failed, skipped, blocked]
        errorMessage:
          type: string
          nullable: true
        stackTrace:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
          description: Duration in milliseconds
        retries:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - testRunId
        - testCaseId
        - testCaseTitle
        - status
        - retries
        - createdAt
        - updatedAt
...