# lib/api-zod/src/generated/types/createTestRunBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Test Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTestRunBody {
  name: string;
  branch?: string;
  commitHash?: string;
  commitMessage?: string;
  triggeredBy?: string;
  ciProvider?: string;
  testCaseIds?: number[];
}